+++ test/karma.source-root-function.js
@@ -3,7 +3,7 @@ const sharedConfig = require('./karma.shared');
 module.exports = function (config) {
   config.set(Object.assign({}, sharedConfig(config, 'test-source-root'), {
     sourceMapLoader: {
-      useSourceRoot(file) {
+      useSourceRoot(_file) {
         return '/sources';
       }
     }