+++ test/karma.source-root-value.js
@@ -4,7 +4,7 @@ module.exports = function (config) {
   config.set(Object.assign({}, sharedConfig(config, 'test-source-root'), {
     sourceMapLoader: {
       useSourceRoot: '/sources',
-      onlyWithURL: true // Just to complete the code coverage
+      onlyWithURL: false // Just to complete the code coverage
     }
   }));
 };