+++ test/src/source.js
@@ -1,2 +1,3 @@
 import shared from './shared.js'
+
 shared()